feat: Aircraft controller

+++ gofly-api/src/main/java/fr/gofly/service/AircraftService.java
@@ -1,11 +1,14 @@
 package fr.gofly.service;
 
+import fr.gofly.helper.AircraftHelper;
+import fr.gofly.helper.UserHelper;
 import fr.gofly.model.Aircraft;
 import fr.gofly.model.User;
 import fr.gofly.repository.AircraftRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
@@ -14,25 +17,45 @@ import java.util.Set;
 public class AircraftService {
 
     private final AircraftRepository aircraftRepository;
+    private final AircraftHelper aircraftHelper;
+    private final UserHelper userHelper;
 
-    public Aircraft createAircraft(Aircraft aircraft) {
-        return aircraftRepository.save(aircraft);
+    public Optional<Aircraft> createAircraft(Aircraft aircraft) {
+        if (aircraftHelper.isMissingMandatoryField(aircraft)) {
+            return Optional.empty();
+        }
+        return Optional.of(aircraftRepository.save(aircraft));
     }
 
-    public Aircraft updateAircraft(Aircraft aircraft) {
-        return aircraftRepository.save(aircraft);
+    public Optional<Aircraft> updateAircraft(Aircraft aircraft, User user) {
+        if (aircraftHelper.isMissingMandatoryField(aircraft) || !aircraftHelper.isAircraftOwnedByUser(aircraft, user) || !userHelper.isAdmin(user)){
+            return Optional.empty();
+        }
+        return Optional.of(aircraftRepository.save(aircraft));
     }
 
-    public void deleteAircraft(Aircraft aircraft) {
-        aircraftRepository.delete(aircraft);
+    public boolean deleteAircraft(Aircraft aircraft, User user) {
+        if (aircraftHelper.isAircraftOwnedByUser(aircraft, user) || userHelper.isAdmin(user)) {
+            aircraftRepository.delete(aircraft);
+            return true;
+        }
+        return false;
     }
 
-    public Optional<Aircraft> getAircraft(Long aircraftId) {
-        return aircraftRepository.findById(aircraftId);
+    public Optional<Aircraft> getAircraft(Integer aircraftId, User user) {
+        Optional<Aircraft> aircraftOptional = aircraftRepository.findById(aircraftId);
+        if (aircraftOptional.isPresent() && aircraftHelper.isAircraftOwnedByUser(aircraftOptional.get(), user) || userHelper.isAdmin(user)) {
+            return aircraftOptional;
+        }
+        return Optional.empty();
     }
 
-    public Optional<Set<Aircraft>> getUserAircraftList(User user) {
+    public Set<Aircraft> getUserAircrafts(User user) {
         return aircraftRepository.findAllByUser(user);
+    }
+
+    public List<Aircraft> getAllAircrafts(User user) {
+        return aircraftRepository.findAll();
     }
 
 }
